Long long long

+++ src/coordinates.c
@@ -7,11 +7,11 @@
 
 int use_eKS_internal;
 int metric;
-double a, hslope; // mks
-double poly_norm, poly_xt, poly_alpha, mks_smooth; // fmks
-double mks3R0, mks3H0, mks3MY1, mks3MY2, mks3MP0; // mks3
-double startx[NDIM], stopx[NDIM], dx[NDIM];
-double R0, Rin, Rout, Rh;
+REAL a, hslope; // mks
+REAL poly_norm, poly_xt, poly_alpha, mks_smooth; // fmks
+REAL mks3R0, mks3H0, mks3MY1, mks3MY2, mks3MP0; // mks3
+REAL startx[NDIM], stopx[NDIM], dx[NDIM];
+REAL R0, Rin, Rout, Rh;
 
 /*
  * Despite the name, this returns r, th coordinates for a KS or BL
@@ -19,7 +19,7 @@ double R0, Rin, Rout, Rh;
  * native coordinates X
  * If METRIC_MINKOWSKI is set, returns spherical coordinates instead
  */
-void bl_coord(double X[NDIM], double *r, double *th)
+void bl_coord(REAL X[NDIM], REAL *r, REAL *th)
 {
 
   if (metric == METRIC_MINKOWSKI) {
@@ -33,7 +33,7 @@ void bl_coord(double X[NDIM], double *r, double *th)
   if (use_eKS_internal) {
     *th = M_PI * X[2];
   } else {
-    double y, thG, thJ;
+    REAL y, thG, thJ;
     switch (metric) {
       case METRIC_MKS:
         *th = M_PI * X[2] + ((1. - hslope) / 2.) * sin(2. * M_PI * X[2]);
@@ -65,12 +65,12 @@ void bl_coord(double X[NDIM], double *r, double *th)
   }
 }
 
-void bl_to_ks(double X[NDIM], double ucon_bl[NDIM], double ucon_ks[NDIM])
+void bl_to_ks(REAL X[NDIM], REAL ucon_bl[NDIM], REAL ucon_ks[NDIM])
 {
-  double r, th;
+  REAL r, th;
   bl_coord(X, &r, &th);
 
-  double trans[NDIM][NDIM];
+  REAL trans[NDIM][NDIM];
   MUNULOOP
     trans[mu][nu] = delta(mu, nu);
 
@@ -83,12 +83,12 @@ void bl_to_ks(double X[NDIM], double ucon_bl[NDIM], double ucon_ks[NDIM])
     ucon_ks[mu] += trans[mu][nu] * ucon_bl[nu];
 }
 
-void ks_to_bl(double X[NDIM], double ucon_ks[NDIM], double ucon_bl[NDIM])
+void ks_to_bl(REAL X[NDIM], REAL ucon_ks[NDIM], REAL ucon_bl[NDIM])
 {
-  double r, th;
+  REAL r, th;
   bl_coord(X, &r, &th);
 
-  double trans[NDIM][NDIM], rev_trans[NDIM][NDIM];
+  REAL trans[NDIM][NDIM], rev_trans[NDIM][NDIM];
   MUNULOOP
     trans[mu][nu] = delta(mu, nu);
 
@@ -106,9 +106,9 @@ void ks_to_bl(double X[NDIM], double ucon_ks[NDIM], double ucon_bl[NDIM])
 /*
  * returns g_{munu} at location specified by X
  */
-void gcov_func(double X[NDIM], double gcov[NDIM][NDIM])
+void gcov_func(REAL X[NDIM], REAL gcov[NDIM][NDIM])
 {
-  double r, th;
+  REAL r, th;
   bl_coord(X, &r, &th);
 
   if (metric == METRIC_MINKOWSKI) {
@@ -121,11 +121,11 @@ void gcov_func(double X[NDIM], double gcov[NDIM][NDIM])
   }
 
   // compute ks metric
-  double Gcov_ks[NDIM][NDIM];
+  REAL Gcov_ks[NDIM][NDIM];
   gcov_ks(r, th, Gcov_ks);
 
   // convert from ks metric to mks/mmks
-  double dxdX[NDIM][NDIM];
+  REAL dxdX[NDIM][NDIM];
   set_dxdX(X, dxdX);
 
   MUNULOOP
@@ -140,13 +140,13 @@ void gcov_func(double X[NDIM], double gcov[NDIM][NDIM])
 }
 
 // compute KS metric at point (r,th) in KS coordinates (cyclic in t, ph)
-inline void gcov_ks(double r, double th, double gcov[NDIM][NDIM])
+inline void gcov_ks(REAL r, REAL th, REAL gcov[NDIM][NDIM])
 {
-  double cth = cos(th);
-  double sth = sin(th);
+  REAL cth = cos(th);
+  REAL sth = sin(th);
 
-  double s2 = sth * sth;
-  double rho2 = r * r + a * a * cth * cth;
+  REAL s2 = sth * sth;
+  REAL rho2 = r * r + a * a * cth * cth;
 
   MUNULOOP gcov[mu][nu] = 0.;
   // Compute KS metric from KS coordinates (cyclic in t,phi)
@@ -165,9 +165,9 @@ inline void gcov_ks(double r, double th, double gcov[NDIM][NDIM])
   gcov[3][3] = s2 * (rho2 + a * a * s2 * (1. + 2. * r / rho2));
 }
 
-inline void gcov_bl(double r, double th, double gcov[NDIM][NDIM])
+inline void gcov_bl(REAL r, REAL th, REAL gcov[NDIM][NDIM])
 {
-  double sth, cth, s2, a2, r2, DD, mu;
+  REAL sth, cth, s2, a2, r2, DD, mu;
   sth = fabs(sin(th));
   s2 = sth * sth;
   cth = cos(th);
@@ -187,7 +187,7 @@ inline void gcov_bl(double r, double th, double gcov[NDIM][NDIM])
 
 }
 
-void set_dxdX(double X[NDIM], double dxdX[NDIM][NDIM])
+void set_dxdX(REAL X[NDIM], REAL dxdX[NDIM][NDIM])
 {
   // Jacobian with respect to KS basis where X is given in
   // non-KS basis
@@ -268,22 +268,22 @@ void set_dxdX(double X[NDIM], double dxdX[NDIM][NDIM])
   }
 }
 
-void set_dXdx(double X[NDIM], double dXdx[NDIM][NDIM]) {
-  double dxdX[NDIM][NDIM];
+void set_dXdx(REAL X[NDIM], REAL dXdx[NDIM][NDIM]) {
+  REAL dxdX[NDIM][NDIM];
   set_dxdX(X, dxdX);
   invert_matrix(dxdX, dXdx);
 }
 
-void vec_to_ks(double X[NDIM], double v_nat[NDIM], double v_ks[NDIM]) {
-  double dxdX[NDIM][NDIM];
+void vec_to_ks(REAL X[NDIM], REAL v_nat[NDIM], REAL v_ks[NDIM]) {
+  REAL dxdX[NDIM][NDIM];
   set_dxdX(X, dxdX);
 
   MULOOP v_ks[mu] = 0.;
   MUNULOOP v_ks[mu] += dxdX[mu][nu] * v_nat[nu];
 }
 
-void vec_from_ks(double X[NDIM], double v_ks[NDIM], double v_nat[NDIM]) {
-  double dXdx[NDIM][NDIM];
+void vec_from_ks(REAL X[NDIM], REAL v_ks[NDIM], REAL v_nat[NDIM]) {
+  REAL dXdx[NDIM][NDIM];
   set_dXdx(X, dXdx);
 
   MULOOP v_nat[mu] = 0.;
@@ -293,11 +293,11 @@ void vec_from_ks(double X[NDIM], double v_ks[NDIM], double v_nat[NDIM]) {
 /*
  * Translate the input camera angles into a canonical Xcam in native coordinates
  */
-void native_coord(double r, double th, double phi, double X[NDIM]) {
+void native_coord(REAL r, REAL th, REAL phi, REAL X[NDIM]) {
   if (metric == METRIC_MINKOWSKI) {
     X[0] = 1; X[1] = r; X[2] = th/180*M_PI; X[3] = phi/180*M_PI;
   } else {
-    double x[NDIM] = {0., r, th/180.*M_PI, phi/180.*M_PI};
+    REAL x[NDIM] = {0., r, th/180.*M_PI, phi/180.*M_PI};
     X[0] = 0.0;
     X[1] = log(r);
     X[2] = root_find(x);
@@ -310,12 +310,12 @@ void native_coord(double r, double th, double phi, double X[NDIM]) {
  * 
  * TODO switch this to native_coord func above...
  */
-double root_find(double X[NDIM])
+REAL root_find(REAL X[NDIM])
 {
-  double th = X[2];
-  double tha, thb, thc;
+  REAL th = X[2];
+  REAL tha, thb, thc;
 
-  double Xa[NDIM], Xb[NDIM], Xc[NDIM];
+  REAL Xa[NDIM], Xb[NDIM], Xc[NDIM];
   Xa[1] = log(X[1]);
   Xa[3] = X[3];
   Xb[1] = Xa[1];
@@ -331,7 +331,7 @@ double root_find(double X[NDIM])
     Xb[2] = stopx[2];
   }
 
-  double tol = 1.e-9;
+  REAL tol = 1.e-9;
   tha = theta_func(Xa);
   thb = theta_func(Xb);
 
@@ -352,7 +352,7 @@ double root_find(double X[NDIM])
     else
       Xb[2] = Xc[2];
 
-    double err = theta_func(Xc) - th;
+    REAL err = theta_func(Xc) - th;
     if (fabs(err) < tol)
       break;
   }
