+++ src/image.h
@@ -8,9 +8,11 @@
 #ifndef IMAGE_H
 #define IMAGE_H
 
+#include "decs.h"
+
 /* imaging */
-void make_ppm(double p[], int nx, int ny, double freq, char filename[]);
-void rainbow_palette(double data, double min, double max, int *pRed,
+void make_ppm(REAL p[], int nx, int ny, REAL freq, char filename[]);
+void rainbow_palette(REAL data, REAL min, REAL max, int *pRed,
                      int *pGreen, int *pBlue);
 
 #endif // IMAGE_H
