+++ src/model_geodesics.h
@@ -11,13 +11,13 @@
 #include "decs.h"
 #include "par.h"
 
-int trace_geodesic(double Xi[NDIM], double Kconi[NDIM], struct of_traj *traj, double eps, int step_max);
-void init_XK(long int i, long int j, int nx, int ny, double Xcam[NDIM],
-             Params params, double fovx, double fovy,
-             double X[NDIM], double Kcon[NDIM]);
+int trace_geodesic(REAL Xi[NDIM], REAL Kconi[NDIM], struct of_traj *traj, REAL eps, int step_max);
+void init_XK(long int i, long int j, int nx, int ny, REAL Xcam[NDIM],
+             Params params, REAL fovx, REAL fovy,
+             REAL X[NDIM], REAL Kcon[NDIM]);
 
 // Internal utilities still used for slow light
-int stop_backward_integration(double X[NDIM], double Xhalf[NDIM], double Kcon[NDIM]);
-double stepsize(double X[NDIM], double K[NDIM], double eps);
+int stop_backward_integration(REAL X[NDIM], REAL Xhalf[NDIM], REAL Kcon[NDIM]);
+REAL stepsize(REAL X[NDIM], REAL K[NDIM], REAL eps);
 
 #endif /* SRC_MODEL_GEODESICS_H_ */
